+++ modules/__init__.py
@@ -1,6 +1,4 @@
-# -*- coding: utf-8 -*-
-
 from modules.banner import *
-from modules.database import DataBase
+from modules.database import DataConnect
 from modules.menu import Menu
 from modules.main import Manager