+++ modules/database.py
@@ -1,18 +1,14 @@
-# -*- coding: utf-8 -*-
 from modules import *
 
-import sqlite3
 import random
 import string
-
-from time import sleep
 
 from base64 import b64encode, b64decode
 from Crypto.Cipher import AES
 from Crypto.Util.Padding import pad, unpad
 
 
-class DataBase:
+class DataConnect:
     """
     DataBase is the class which contains database SQlite
     functions and encryption using AES and base 64.
@@ -25,10 +21,11 @@ class DataBase:
         master_pw [str] -- master password
     """
 
-    def __init__(self, obj, master_pssw: str) -> None:
-        self.obj_ = obj
+    def __init__(self, master_pssw: str, obj) -> None:
         # Connect with the SQlite and create the table.
-        self.obj_.cursor.execute(
+        self.master_pw = master_pssw
+        self.obj_ = obj
+        self.query_command(
             """CREATE TABLE IF NOT EXISTS passwords (
                 id INTEGER PRIMARY KEY,
                 platform TEXT NOT NULL,
@@ -37,8 +34,77 @@ class DataBase:
                 url TEXT NOT NULL
             );"""
         )
-        self.master_pw = master_pssw
         self.specialchars = "_!@#$%&*()-"
+    
+    def query_command(self, sql: str, *args) -> None:
+        """
+        Query commands 
+
+        Arguments
+            sql [str] -- sqlite command
+        """
+        return self.obj_.cur.execute(sql, *args)
+    
+    def select_all(self, table: str) -> list:
+        """
+        Select all 
+
+        Arguments 
+            table [str] -- the table to be selected
+            star [str] -- what select from the table
+        """
+        return self.query_command(f"SELECT * FROM {table}")
+    
+    def update_where(self, value: str, new: str, id_opt: str) -> None:
+        """
+        Update somewhere
+
+        Arguments 
+            value [str] -- the value to update
+            new [str] -- new data to be informed
+            id_opt [str] -- the id to be selected
+        """
+        cmd = self.query_command(f"UPDATE passwords SET {value} = '{new}' WHERE id = '{id_opt}'")
+        self.obj_.conn.commit()
+        
+        return cmd
+        
+    def delete_where(self, id_opt: str) -> None:
+        """
+        Delete somewhere 
+
+        Arguments 
+            id_opt [str] -- the id to be deleted
+        """
+        cmd = self.query_command(f"DELETE FROM passwords WHERE id = '{id_opt}'")
+        self.obj_.conn.commit()
+
+        return cmd
+
+    def drop_table(self, table: str) -> None: 
+        """
+        Drop tables
+
+        Arguments 
+            table [str] -- table to be deleted
+        """
+        self.query_command(f"DROP TABLE WHERE {table}")
+        self.obj_.conn.commit()
+    
+    def delete_all_data(self):
+        self.drop_table("passwords")
+        self.drop_table("masterpassword")
+
+    def verify_id(self, id_opt: str):
+        """
+        Verify if the ID is correct 
+
+        Arguments
+            id_opt [str] -- the id informed
+        """
+        id_list = [row[0] for row in self.select_all('passwords')]
+        if id_opt not in str(id_list): 
+            return print(f"[red]{self.obj_.xmark_} The ID is not correct[/]")
 
     def encryption(self, pssw: str, key: str) -> tuple:
         """
@@ -91,7 +157,7 @@ class DataBase:
         pw_len = int(input("What length would you like your password to be? (At least 8) "))
 
         if pw_len < 8:
-            print("\n[red] The password is not long enough. Please try again.[/red]\n")
+            print("\n[red] The password is not long enough. Please try again.[/]\n")
             return self.generate_password()
         else:
             # Generating a password
@@ -118,13 +184,12 @@ class DataBase:
         Returns 
             [tuple] -- Sensitives datas saved in the SQLite.
         """
-        self.obj_.cursor.execute("SELECT id FROM passwords;")
         # The ID must be the length of the datas stored in the database plus one
         # then the user can change/delete informations.
-        id = len(self.obj_.cursor.fetchall())
+        id = len(self.query_command("SELECT id FROM passwords;").fetchall())
 
         while True:
-            if id in self.obj_.cursor.execute("SELECT id FROM passwords;"):
+            if id in self.query_command("SELECT id FROM passwords;"):
                 # Verify if the ID exist in the database
                 # if exist the code will add one more
                 id += 1
@@ -139,8 +204,8 @@ class DataBase:
             concatenate = initial_value + "|" + contatenate
             infos.append(concatenate)
         # Insert each value in the table passwords
-        self.obj_.cursor.execute(f"""INSERT INTO passwords VALUES('{id}', '{infos[0]}', '{infos[1]}', '{infos[2]}', '{infos[3]}')""")
-        self.obj_.datab.commit()
+        self.query_command(f"INSERT INTO passwords VALUES('{id}', '{infos[0]}', '{infos[1]}', '{infos[2]}', '{infos[3]}')")
+        self.obj_.conn.commit()
 
         print(f"[green]\n{self.obj_.checkmark_} Thank you! Datas were successfully added.[/green]")
 
@@ -156,23 +221,13 @@ class DataBase:
         Returns
             [str] The new password changed in the database
         """
-        self.obj_.cursor.execute("SELECT COUNT(*) from passwords;")
-        if self.obj_.cursor.fetchall()[0][0] == 0: 
-            # verify if the database is empty - cannot opperate in a empty database
-            print(f"[red]{self.obj_.xmark_} The database is empty. Try adding a password.[/red]")
-
-        else:
-            id_list = [row[0] for row in self.obj_.cursor.execute("SELECT * FROM passwords;")]
-            if id_opt not in str(id_list): 
-                return print(f"[red]{self.obj_.xmark_} The ID is not correct[/red]")
-            
-            initial_value, concatenate = self.encryption(new, self.master_pw)
-            ct_new_info = initial_value + "|" + concatenate
-            
-            self.obj_.cursor.execute(f"""UPDATE passwords SET {option} = '{ct_new_info}' WHERE id = '{id_opt}'""")
-            self.obj_.datab.commit()
-
-            print(f"[green]{self.obj_.checkmark_} The {option} of the ID {id_opt} has successfully changed to {new}.[/green]")
+        self.verify_id(id_opt)
+        
+        initial_value, concatenate = self.encryption(new, self.master_pw)
+        ct_new_info = initial_value + "|" + concatenate
+        
+        self.update_where(option, ct_new_info, id_opt)
+        print(f"[green]{self.obj_.checkmark_} The {option} of the ID {id_opt} has successfully changed to {new}.[/green]")
 
     def look_up(self, id_opt: str) -> None:
         """
@@ -184,18 +239,9 @@ class DataBase:
         Returns
             [tuple] The passwords stored
         """
-
-        self.obj_.cursor.execute("SELECT COUNT(*) from passwords;")
-        if self.obj_.cursor.fetchall()[0][0] == 0:
-            # verify if the database is empty - cannot opperate in a empty database
-            print(f"[red]{self.obj_.xmark_} The database is empty. Try adding a password.[/red]")
-
-        id_list = [row[0] for row in self.obj_.cursor.execute("SELECT * FROM passwords;")]
-        if id_opt not in str(id_list): 
-            return print(f"[red]{self.obj_.xmark_} The ID is not correct[/red]")
-            
+        self.verify_id(id_opt)
         infos = []
-        for row in self.obj_.cursor.execute(f"SELECT * FROM passwords WHERE id = '{id_opt}';"):
+        for row in self.query_command(f"SELECT * FROM passwords WHERE id = '{id_opt}';"):
             infos.append(row[1])
             infos.append(row[2])
             infos.append(row[3])
@@ -211,7 +257,8 @@ class DataBase:
             ]
             
             infos.clear()
-            return (f"\n[yellow][ID: {row[0]}] {decrypted[0].decode()}[/yellow]\n"
+            return (
+                f"\n[yellow][ID: {row[0]}] {decrypted[0].decode()}[/yellow]\n"
                 f"[green]Email: {decrypted[1].decode()}\n"
                 f"Password: {decrypted[2].decode()}\n"
                 f"URL: {decrypted[3].decode()}[/green]\n"
@@ -224,14 +271,13 @@ class DataBase:
         Returns 
             [str] -- List of passwords
         """
-        self.obj_.cursor.execute("SELECT COUNT(*) from passwords;")
-        if self.obj_.cursor.fetchall()[0][0] == 0: 
+        if self.query_command("SELECT COUNT(*) from passwords;").fetchall()[0][0] == 0:
             # verify if the database is empty - cannot opperate in a empty database
             raise PermissionError
         
         print('[yellow]Current passwords stored:[/yellow]')
         infos = []
-        for row in self.obj_.cursor.execute("SELECT * FROM passwords;"):
+        for row in self.select_all('passwords'):
             infos.append(row[1])
             infos.append(row[2])
 
@@ -245,59 +291,36 @@ class DataBase:
             ]
             
             infos.clear()
-
             print(f"[yellow][ID: {row[0]}] Platform: {decrypted[0].decode()}[/yellow]")
         
-    def delete_normal(self, id_opt: str) -> None:
+    def delete_one_password(self, id_opt: str) -> None:
         """Delete one password
 
         Arguments 
             id_opt [str] -- the ID chosed
         """
-        self.obj_.cursor.execute("SELECT COUNT(*) from passwords;")
-        if self.obj_.cursor.fetchall()[0][0] == 0: 
-            # verify if the database is empty - cannot opperate in a empty database
-            print(f"[red]{self.obj_.xmark_} The database is empty. Try adding a password.[/red]")
+        self.verify_id(id_opt)
 
-        else:
-            id_list = [row[0] for row in self.obj_.cursor.execute("SELECT * FROM passwords;")]
-            if id_opt not in str(id_list): 
-                return print(f"[red]{self.obj_.xmark_} The ID is not correct[/red]")
+        self.delete_where(id_opt)
+        print(f"[green]\n{self.obj_.checkmark_} The password was successfully deleted.\n[/green]")
 
-            self.obj_.cursor.execute(f"DELETE from passwords WHERE id = '{id_opt}'") 
-            self.obj_.datab.commit()
-            print(f"[green]\n{self.obj_.checkmark_} The password was successfully deleted.\n[/green]")
-
-    def delete_master(self):
-        """
-        Delete master password all the data
-        """
-        print("[red]Deleting all the passwords...[/red]")
-        self.obj_.cursor.execute("DROP TABLE passwords;")
-        self.obj_.cursor.execute("DROP TABLE masterpassword;")
-        self.obj_.datab.commit()
-
-    def delete_all(self) -> None:
+    def delete_all_passwords(self, entered_master: str) -> None:
         """
         Delete all passwords stored in the database 
         (Normal passwords not the master)
 
+        Arguments
+            entered_master [str] -- master password to verify 
+
         Returns
             DataBase empty (SQlite)
         """
-        self.obj_.cursor.execute("SELECT COUNT(*) from passwords;")
-        if self.obj_.cursor.fetchall()[0][0] == 0: 
+        if self.query_command("SELECT COUNT(*) from passwords;").fetchall()[0][0] == 0: 
             # verify if the database is empty - cannot opperate in a empty database
-            print(f"[red]{self.obj_.xmark_} The database is empty. Try adding a password.[/red]")
+            print(f"[red]{self.obj_.xmark_} The database is empty. Try adding a password.[/]")
 
-        else:
-            print("[red]Deleting all data... (database)[/red]")
-            print("Removing...")
-            sleep(2)
-
-            # Dropping the datable
-            self.obj_.cursor.execute("DROP TABLE passwords;")
-            self.obj_.datab.commit()
-
-            sleep(1)
-            print(f"[green]{self.obj_.checkmark_} Done. All the passwords stored had been deleted with success.[/green]")
+        if self.master_pw == entered_master:
+            self.drop_table('passwords')
+        else: 
+            return print(f'[red]{self.obj_.xmark_} The master password is not correct.[/]')
+